new package: face_detection_msgs
This package contains definitions for various custom message types
useful for face detection. Specifically it creates a custom ``Face``
message type which contains a bounding box (or "region of interest") for
a face and a ``Faces`` message which contains an array of ``Face``
messages.

Note that this is the first package in our stack which requires
building. If you are using the Docker-based workflow, you can simply
build the stack via ``make build``. (Or ``make`` since ``build`` is the
default target.)